requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.15
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-qdrant==0.2.1
langchain-text-splitters==0.3.11
qdrant-client==1.15.1
pymupdf
pandas
jupyter
Pillow
tiktoken
rapidocr-onnxruntime
sentence-transformers==5.1.1
streamlit
requests